feat: improve page speed with better code splitting, more efficient rendering of offer details
