+++ src/pages/favorites/index.tsx
@@ -1,72 +1,14 @@
-import { useMemo } from "react";
-import { useSearchParams } from "react-router-dom";
-import BackButton from "../../components/buttons/back-button";
-import FilterButton from "../../components/buttons/filter-button";
-import SortButton from "../../components/buttons/sort-button";
-import { InfoBox } from "../../components/info-box/info-box";
-import OfferDetail from "../../components/offer/offer-detail";
-import { offers } from "../../content/content";
-import { useLanguage } from "../../hooks/use-language";
-import { useI18n } from "../../i18n/use-i18n";
+import { lazy, Suspense } from "react";
 import { Layout } from "../../layout/layout";
-import { useFavoritesStore } from "../../store/favorites-store";
+
+const LazyFavorites = lazy(() => import("../../components/lazy/favorites"));
 
 export default function Index() {
-	const language = useLanguage();
-	const i18n = useI18n(language);
-
-	const [searchParams] = useSearchParams();
-	const sortAscending = searchParams.get("sort") === "asc";
-
-	const favorites = useFavoritesStore((state) => state.favorites);
-
-	const filteredOffers = useMemo(() => {
-		const filtered = offers
-			.filter((offer) => favorites.includes(offer.slug))
-			.filter((offer) => offer.language === language);
-		const sorted = filtered.sort((a, b) => {
-			if (sortAscending) {
-				return a.provider.localeCompare(b.provider);
-			}
-			return b.provider.localeCompare(a.provider);
-		});
-
-		return sorted;
-	}, [offers, favorites, sortAscending]);
-
 	return (
 		<Layout>
-			<div>
-				<div
-					className={`w-full bg-primary-blue flex flex-row justify-center items-center text-[#ffffff] p-3 mb-10 font-bold text-xl`}
-				>
-					{i18n["menuItem.favorits"]}
-				</div>
-
-				<div className="max-w-3xl mx-auto flex flex-col">
-					<div className="flex flex-row w-full justify-between mb-6 px-4 lg:px-0">
-						<SortButton></SortButton>
-						<FilterButton></FilterButton>
-					</div>
-
-					<div className="mx-4 lg:mx-0 flex flex-row items-center gap-2 py-3">
-						<p className="text-md text-berlin-green">
-							{filteredOffers.length} {i18n["allOffers.offerFound"]}
-						</p>
-					</div>
-
-					<div className="w-full border-b border-separator mb-5"></div>
-					<div className="flex flex-col pt-4 mb-5">
-						{filteredOffers.map((offer, idx) => (
-							<OfferDetail offer={offer} key={`${idx}-${offer.provider}`} />
-						))}
-					</div>
-					<div className="mb-10 px-5 lg:px-0">
-						<BackButton title={i18n["return"]}></BackButton>
-					</div>
-				</div>
-			</div>
-			<InfoBox showProviderHint={false} maxWidth="max-w-3xl"></InfoBox>
+			<Suspense fallback={<div className="w-svw h-svh" />}>
+				<LazyFavorites />
+			</Suspense>
 		</Layout>
 	);
 }
